Improve robustness and variable types

+++ gol.h
@@ -5,7 +5,7 @@ struct universe {
     int columns; // Number of columns
     int generation; // Number indicating which generation the universe is (beginning at 0)
     int numAlive; // Number of alive cells in current generation
-    float prevAvgAlive; // Average number of cells alive across all previous generations
+    int prevAvgAlive; // Average number of cells alive across all previous generations
 };
 
 /*Do not modify the next seven lines*/
